Added DM capabilities
Added a Host game object with DM specific code and controls, using a state system. Added minor animation to opening of DMs monster drawer. New bug: spawning a wall will not mark the appropriate cell as "unwalkable". Think this may be the result of when Pathfinding is instantiated. Preferably make this happen before the scene change AND/OR a server stored object.